Update mc
Added timer killer if user terminates Minecraft before time expires on it's own. i.e.; Previous to this addition if a user terminated Minecraft the terminal would still be running the 'mchammer' function until either 9pm or midnight as it was unaware that the game had been exited... Now the terminal prompt will be returned up to 10 seconds after the game is quit.

+++ mc
@@ -26,7 +26,7 @@ cyear=$(date +%Y)
 # because I didn't want it running crazy and I didn't want to wait any longer to kill MC.
 # You can change the value of sleep, but it's just gonna take Johnny that much longer to kill Herobrine
 # After the time runs out. i.e.; There will be a Delay before the game actually ends. He's not punctual
-# and the higher you set the sleep value the later he will sleep in.
+# and the higher you set the sleep value the later he will sleep in... So I do not reccomend changing this.
 
 mchammer ()
 {
@@ -35,6 +35,12 @@ mchammer ()
   do
   clh=$(date +%H)
   sleep 10s
+ jpid=$(pgrep java)
+ if [ -z "$jpid" ]
+  then 
+  exit
+ fi  
+  
 done
 kill $(pgrep java)
 }
